WEEK_4/requirements.txt
bcrypt==3.2.0
cffi==1.15.0
click==8.0.1
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.2
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Pillow==8.4.0
pycparser==2.21
six==1.16.0
SQLAlchemy==1.4.28
Werkzeug==2.0.1
WTForms==2.3.3
